meal_mind_streamlit/requirements.txt
streamlit
snowflake-connector-python
snowflake-snowpark-python
python-dotenv
pandas
langchain
langchain-community
langgraph
pydantic
langchain-snowflake
requests